first 98 Fibonacci numbers, starting with 1 and 2

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -6,7 +6,7 @@
   * Return: Always 0 (Success)
   */
 
-unsigned long main(void)
+int main(void)
 {
 	int num;
 	int i;
